Portfolio, Project, ProjectCard, Projects, ProjectsCount: moves to src/components/portfolio; index: updates import paths accordingly
